refactor(employee-form): unused variable removal

+++ src/app/Employee/employee-form/employee-table/employee-table.component.ts
@@ -1,6 +1,5 @@
 import { HttpParams } from '@angular/common/http';
 import { Component, OnInit } from '@angular/core';
-import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { CountryService } from '../../../curd.service';
 @Component({
   selector: 'app-employee-table',
@@ -11,29 +10,27 @@ import { CountryService } from '../../../curd.service';
 export class EmployeeTableComponent implements OnInit {
   error: null;
   employeeTable: any ;
-  content: any
-  bookForm: FormGroup;
   selectedItem: any;
   titleName: string;
   submitted: boolean;
   selectedIndex: number;
-  isConfirmationOpen:boolean;
+  employeeFormOpen:boolean;
   params: HttpParams;
-  isOpen:boolean;
+  isConfirmationOpen:boolean;
 
-  constructor(private formBuilder: FormBuilder,private dataservice: CountryService) {}
+  constructor(private dataservice: CountryService) {}
 
   onAdd() {
     this.selectedItem = null;
     this.titleName = 'Add Employee';
-    this.isConfirmationOpen=true;
+    this.employeeFormOpen=true;
   }
 
   onEdit(item:any) {
     this.selectedItem = item;
     this.selectedIndex = item.id;
     this.titleName = 'Edit Employee';
-    this.isConfirmationOpen=true;
+    this.employeeFormOpen=true;
   }
 
   onStatusChanged(event: any) {
@@ -46,17 +43,18 @@ export class EmployeeTableComponent implements OnInit {
     this.getEmployee();
   }
 
-  delete(id:number) {
-    this.dataservice.deleteEmployee(id).subscribe((data): void => {
-      this.getEmployee();
-
-    });
+  delete(id:number,employee:string) {
+    if (confirm('Are you sure to delete ' + employee)) {
+      this.dataservice.deleteEmployee(id).subscribe((data): void => {
+        this.getEmployee();
+      });
+    }
   }
 
   getId(item:any) {
     this.selectedIndex = item.id;
     console.log(this.selectedIndex);
-    this.isOpen=true;
+    this.isConfirmationOpen=true;
   }
 
   ngOnInit(): void {
@@ -85,13 +83,13 @@ export class EmployeeTableComponent implements OnInit {
         this.error = error;
         throw error;
       });
-      this.isConfirmationOpen=false;
+      this.employeeFormOpen=false;
     } else {
       this.dataservice.addPostEmployee(value).subscribe(responseData => {
         console.log(responseData);
         this.getEmployee();
       });
-      this.isConfirmationOpen=false;
+      this.employeeFormOpen=false;
     }
   }
